Added: Fixes version 0.5.5

+++ src/intf/strings.h
@@ -7,4 +7,4 @@
 int strcmp(const char* s1, const char* s2);
 int strncmp(const char* s1, const char* s2, size_t n);
 int strlen(const char* str);
-void* memcpy(void* dest, const void* src, int count);
+void* memcpy(void* dest, const void* src, size_t count);
